requirements.txt
aiogram>=2.20
requests>=2.25.1
Pillow>=9.5.0
